修正 shmop / shmget 参数类型 (int -> size_t)

+++ include/ytht/shmop.h
@@ -1,6 +1,6 @@
 #ifndef BMYBBS_SHMOP_H
 #define BMYBBS_SHMOP_H
-void *try_get_shm(int key, int size, int flag);
+void *try_get_shm(int key, size_t size, int flag);
 void *get_shm(int key, int size);
 
 #define get_old_shm(x,y) try_get_shm(x,y,0)
